Refactor app shell and thread state modules
