+++ src/composables/useDesktopState.ts
@@ -56,970 +56,85 @@ import type {
   UiThread,
 } from '../types/codex'
 import { normalizePathForUi, toProjectName } from '../pathUtils.js'
+import {
+  GLOBAL_SERVER_REQUEST_SCOPE,
+  MODEL_FALLBACK_ID,
+  NEW_THREAD_COLLABORATION_MODE_CONTEXT,
+  REASONING_EFFORT_OPTIONS,
+  clamp,
+  cloneStringKeyedRecord,
+  createStringKeyedRecord,
+  loadProjectDisplayNames,
+  loadProjectOrder,
+  loadReadStateMap,
+  loadSelectedCollaborationModeMap,
+  loadSelectedModelMap,
+  loadSelectedThreadId,
+  loadThreadScrollStateMap,
+  loadThreadTerminalOpenMap,
+  loadThreadTokenUsageMap,
+  normalizeCollaborationMode,
+  normalizeProviderContextId,
+  normalizeStoredModelId,
+  omitStringKeyedRecordKey,
+  pruneThreadContextStateMap,
+  readSelectedCollaborationMode,
+  readSelectedModel,
+  saveProjectDisplayNames,
+  saveProjectOrder,
+  saveReadStateMap,
+  saveSelectedCollaborationModeMap,
+  saveSelectedModelMap,
+  saveSelectedThreadId,
+  saveThreadScrollStateMap,
+  saveThreadTerminalOpenMap,
+  saveThreadTokenUsageMap,
+  toProviderModelContextId,
+  toThreadContextId,
+  writeSelectedCollaborationModeForContext,
+} from '../features/thread-state/storage'
+import {
+  areGroupArraysEqual,
+  areStringArraysEqual,
+  flattenThreads,
+  mergeIncomingWithLocalInProgressThreads,
+  mergeProjectOrder,
+  mergeThreadGroups,
+  omitKey,
+  omitKeys,
+  orderGroupsByProjectOrder,
+  pruneThreadStateMap,
+  reorderStringArray,
+  toForkedThreadTitle,
+  toOptimisticThreadTitle,
+  toProjectNameFromWorkspaceRoot,
+} from '../features/thread-state/merge'
+import {
+  type TurnActivityState,
+  type TurnCompletedInfo,
+  type TurnErrorState,
+  type TurnStartedInfo,
+  type TurnSummaryState,
+  areMessageArraysEqual,
+  areMessageFieldsEqual,
+  areTurnActivitiesEqual,
+  areTurnSummariesEqual,
+  areUiFileChangesEqual,
+  insertTurnSummaryMessage,
+  isUnsupportedChatGptModelError,
+  mergeMessages,
+  normalizeMessageText,
+  parseIsoTimestamp,
+  removePersistedLiveMessages,
+  removeRedundantLiveAgentMessages,
+  upsertMessage,
+} from '../features/thread-state/messages'
 
-function flattenThreads(groups: UiProjectGroup[]): UiThread[] {
-  return groups.flatMap((group) => group.threads)
-}
-
-const READ_STATE_STORAGE_KEY = 'codex-web-local.thread-read-state.v1'
-const SCROLL_STATE_STORAGE_KEY = 'codex-web-local.thread-scroll-state.v1'
-const THREAD_TOKEN_USAGE_STORAGE_KEY = 'codex-web-local.thread-token-usage.v1'
-const THREAD_TERMINAL_OPEN_STORAGE_KEY = 'codex-web-local.thread-terminal-open.v1'
-const SELECTED_THREAD_STORAGE_KEY = 'codex-web-local.selected-thread-id.v1'
-const SELECTED_MODEL_BY_CONTEXT_STORAGE_KEY = 'codex-web-local.selected-model-by-context.v1'
-const LEGACY_SELECTED_MODEL_STORAGE_KEY = 'codex-web-local.selected-model-id.v1'
-const PROJECT_ORDER_STORAGE_KEY = 'codex-web-local.project-order.v1'
-const PROJECT_DISPLAY_NAME_STORAGE_KEY = 'codex-web-local.project-display-name.v1'
-const COLLABORATION_MODE_STORAGE_KEY = 'codex-web-local.collaboration-mode-by-context.v1'
-const LEGACY_COLLABORATION_MODE_STORAGE_KEY = 'codex-web-local.collaboration-mode.v1'
-const NEW_THREAD_COLLABORATION_MODE_CONTEXT = '__new-thread__'
-const NEW_THREAD_PROVIDER_MODEL_CONTEXT_PREFIX = '__new-thread-provider__::'
 const EVENT_SYNC_DEBOUNCE_MS = 220
 const BACKGROUND_THREAD_PAGINATION_DELAY_MS = 10_000
 const RATE_LIMIT_REFRESH_DEBOUNCE_MS = 500
 const TURN_START_FOLLOW_UP_SYNC_DELAY_MS = 3000
 const RECENT_THREAD_MESSAGE_LOAD_REUSE_MS = 2000
-const REASONING_EFFORT_OPTIONS: ReasoningEffort[] = ['none', 'minimal', 'low', 'medium', 'high', 'xhigh']
-const GLOBAL_SERVER_REQUEST_SCOPE = '__global__'
-const MODEL_FALLBACK_ID = 'gpt-5.2-codex'
-
-function loadReadStateMap(): Record<string, string> {
-  if (typeof window === 'undefined') return {}
-
-  try {
-    const raw = window.localStorage.getItem(READ_STATE_STORAGE_KEY)
-    if (!raw) return {}
-
-    const parsed = JSON.parse(raw) as unknown
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return {}
-    return parsed as Record<string, string>
-  } catch {
-    return {}
-  }
-}
-
-function saveReadStateMap(state: Record<string, string>): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(READ_STATE_STORAGE_KEY, JSON.stringify(state))
-}
-
-function normalizeCollaborationMode(value: unknown): CollaborationModeKind {
-  return value === 'plan' ? 'plan' : 'default'
-}
-
-function normalizeStoredModelId(value: unknown): string {
-  return typeof value === 'string' ? value.trim() : ''
-}
-
-function createStringKeyedRecord<T>(): Record<string, T> {
-  return Object.create(null) as Record<string, T>
-}
-
-function cloneStringKeyedRecord<T>(record: Record<string, T>): Record<string, T> {
-  const next = createStringKeyedRecord<T>()
-  for (const [key, value] of Object.entries(record)) {
-    next[key] = value
-  }
-  return next
-}
-
-function omitStringKeyedRecordKey<T>(record: Record<string, T>, key: string): Record<string, T> {
-  if (!(key in record)) return record
-  const next = createStringKeyedRecord<T>()
-  for (const [entryKey, value] of Object.entries(record)) {
-    if (entryKey !== key) {
-      next[entryKey] = value
-    }
-  }
-  return next
-}
-
-function pruneThreadContextStateMap<T>(
-  stateMap: Record<string, T>,
-  threadIds: Set<string>,
-): Record<string, T> {
-  let changed = false
-  const next = createStringKeyedRecord<T>()
-  for (const [contextId, value] of Object.entries(stateMap)) {
-    if (
-      contextId === NEW_THREAD_COLLABORATION_MODE_CONTEXT
-      || contextId.startsWith(NEW_THREAD_PROVIDER_MODEL_CONTEXT_PREFIX)
-      || threadIds.has(contextId)
-    ) {
-      next[contextId] = value
-      continue
-    }
-    changed = true
-  }
-  return changed ? next : stateMap
-}
-
-function normalizeProviderContextId(providerId: string): string {
-  const normalized = providerId.trim().toLowerCase()
-  return normalized || 'codex'
-}
-
-function toProviderModelContextId(providerId: string): string {
-  const normalizedProviderId = normalizeProviderContextId(providerId)
-  if (!normalizedProviderId) return ''
-  return `${NEW_THREAD_PROVIDER_MODEL_CONTEXT_PREFIX}${normalizedProviderId}`
-}
-
-function toThreadContextId(threadId: string): string {
-  const normalizedThreadId = threadId.trim()
-  return normalizedThreadId || NEW_THREAD_COLLABORATION_MODE_CONTEXT
-}
-
-function loadSelectedModelMap(): Record<string, string> {
-  if (typeof window === 'undefined') return createStringKeyedRecord<string>()
-
-  try {
-    const raw = window.localStorage.getItem(SELECTED_MODEL_BY_CONTEXT_STORAGE_KEY)
-    if (raw) {
-      const parsed = JSON.parse(raw) as unknown
-      if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return createStringKeyedRecord<string>()
-
-      const next = createStringKeyedRecord<string>()
-      for (const [contextId, value] of Object.entries(parsed as Record<string, unknown>)) {
-        if (typeof contextId !== 'string' || contextId.length === 0) continue
-        const normalizedModelId = normalizeStoredModelId(value)
-        if (normalizedModelId) {
-          next[contextId] = normalizedModelId
-        }
-      }
-      return next
-    }
-  } catch {
-    // Fall back to the legacy global preference below.
-  }
-
-  const legacyModelId = normalizeStoredModelId(window.localStorage.getItem(LEGACY_SELECTED_MODEL_STORAGE_KEY))
-  const next = createStringKeyedRecord<string>()
-  if (legacyModelId) {
-    next[NEW_THREAD_COLLABORATION_MODE_CONTEXT] = legacyModelId
-  }
-  return next
-}
-
-function readSelectedModel(
-  state: Record<string, string>,
-  threadId: string,
-): string {
-  const contextId = toThreadContextId(threadId)
-  const contextModelId = normalizeStoredModelId(state[contextId])
-  if (contextModelId) return contextModelId
-  return normalizeStoredModelId(state[NEW_THREAD_COLLABORATION_MODE_CONTEXT])
-}
-
-function saveSelectedModelMap(state: Record<string, string>): void {
-  if (typeof window === 'undefined') return
-  try {
-    if (Object.keys(state).length === 0) {
-      window.localStorage.removeItem(SELECTED_MODEL_BY_CONTEXT_STORAGE_KEY)
-    } else {
-      window.localStorage.setItem(SELECTED_MODEL_BY_CONTEXT_STORAGE_KEY, JSON.stringify(state))
-    }
-    window.localStorage.removeItem(LEGACY_SELECTED_MODEL_STORAGE_KEY)
-  } catch {
-    // Keep in-memory selection working even if localStorage writes fail.
-  }
-}
-
-function loadSelectedCollaborationModeMap(): Record<string, CollaborationModeKind> {
-  if (typeof window === 'undefined') return createStringKeyedRecord<CollaborationModeKind>()
-
-  try {
-    const raw = window.localStorage.getItem(COLLABORATION_MODE_STORAGE_KEY)
-    if (raw) {
-      const parsed = JSON.parse(raw) as unknown
-      if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) {
-        return createStringKeyedRecord<CollaborationModeKind>()
-      }
-
-      const next = createStringKeyedRecord<CollaborationModeKind>()
-      for (const [contextId, value] of Object.entries(parsed as Record<string, unknown>)) {
-        if (typeof contextId !== 'string' || contextId.length === 0) continue
-        const normalizedMode = normalizeCollaborationMode(value)
-        if (normalizedMode === 'plan') {
-          next[contextId] = normalizedMode
-        }
-      }
-      return next
-    }
-  } catch {
-    // Fall back to the legacy global preference below.
-  }
-
-  const legacyMode = normalizeCollaborationMode(window.localStorage.getItem(LEGACY_COLLABORATION_MODE_STORAGE_KEY))
-  const next = createStringKeyedRecord<CollaborationModeKind>()
-  if (legacyMode === 'plan') {
-    next[NEW_THREAD_COLLABORATION_MODE_CONTEXT] = 'plan'
-  }
-  return next
-}
-
-function readSelectedCollaborationMode(
-  state: Record<string, CollaborationModeKind>,
-  threadId: string,
-): CollaborationModeKind {
-  const contextId = toThreadContextId(threadId)
-  return normalizeCollaborationMode(state[contextId])
-}
-
-function writeSelectedCollaborationModeForContext(
-  state: Record<string, CollaborationModeKind>,
-  threadId: string,
-  mode: CollaborationModeKind,
-): Record<string, CollaborationModeKind> {
-  const contextId = toThreadContextId(threadId)
-  if (mode === 'plan') {
-    const next = cloneStringKeyedRecord(state)
-    next[contextId] = 'plan'
-    return next
-  }
-  return omitStringKeyedRecordKey(state, contextId)
-}
-
-function saveSelectedCollaborationModeMap(state: Record<string, CollaborationModeKind>): void {
-  if (typeof window === 'undefined') return
-  try {
-    if (Object.keys(state).length === 0) {
-      window.localStorage.removeItem(COLLABORATION_MODE_STORAGE_KEY)
-    } else {
-      window.localStorage.setItem(COLLABORATION_MODE_STORAGE_KEY, JSON.stringify(state))
-    }
-    window.localStorage.removeItem(LEGACY_COLLABORATION_MODE_STORAGE_KEY)
-  } catch {
-    // Keep in-memory mode selection working even if localStorage writes fail.
-  }
-}
-
-function clamp(value: number, minValue: number, maxValue: number): number {
-  return Math.min(Math.max(value, minValue), maxValue)
-}
-
-function normalizeThreadScrollState(value: unknown): ThreadScrollState | null {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return null
-
-  const rawState = value as Record<string, unknown>
-  if (typeof rawState.scrollTop !== 'number' || !Number.isFinite(rawState.scrollTop)) return null
-  if (typeof rawState.isAtBottom !== 'boolean') return null
-
-  const normalized: ThreadScrollState = {
-    scrollTop: Math.max(0, rawState.scrollTop),
-    isAtBottom: rawState.isAtBottom,
-  }
-
-  if (typeof rawState.scrollRatio === 'number' && Number.isFinite(rawState.scrollRatio)) {
-    normalized.scrollRatio = clamp(rawState.scrollRatio, 0, 1)
-  }
-
-  return normalized
-}
-
-function loadThreadScrollStateMap(): Record<string, ThreadScrollState> {
-  if (typeof window === 'undefined') return {}
-
-  try {
-    const raw = window.localStorage.getItem(SCROLL_STATE_STORAGE_KEY)
-    if (!raw) return {}
-
-    const parsed = JSON.parse(raw) as unknown
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return {}
-
-    const normalizedMap: Record<string, ThreadScrollState> = {}
-    for (const [threadId, state] of Object.entries(parsed as Record<string, unknown>)) {
-      if (!threadId) continue
-      const normalizedState = normalizeThreadScrollState(state)
-      if (normalizedState) {
-        normalizedMap[threadId] = normalizedState
-      }
-    }
-    return normalizedMap
-  } catch {
-    return {}
-  }
-}
-
-function saveThreadScrollStateMap(state: Record<string, ThreadScrollState>): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(SCROLL_STATE_STORAGE_KEY, JSON.stringify(state))
-}
-
-function normalizeStoredTokenCount(value: unknown): number | null {
-  if (typeof value === 'number' && Number.isFinite(value)) {
-    return Math.max(0, Math.trunc(value))
-  }
-
-  if (typeof value === 'string' && value.trim().length > 0) {
-    const parsed = Number(value)
-    if (Number.isFinite(parsed)) {
-      return Math.max(0, Math.trunc(parsed))
-    }
-  }
-
-  return null
-}
-
-function normalizeTokenUsageBreakdown(value: unknown): UiThreadTokenUsage['last'] | null {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return null
-
-  const record = value as Record<string, unknown>
-  return {
-    totalTokens: normalizeStoredTokenCount(record.totalTokens) ?? 0,
-    inputTokens: normalizeStoredTokenCount(record.inputTokens) ?? 0,
-    cachedInputTokens: normalizeStoredTokenCount(record.cachedInputTokens) ?? 0,
-    outputTokens: normalizeStoredTokenCount(record.outputTokens) ?? 0,
-    reasoningOutputTokens: normalizeStoredTokenCount(record.reasoningOutputTokens) ?? 0,
-  }
-}
-
-function normalizeThreadTokenUsage(value: unknown): UiThreadTokenUsage | null {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) return null
-
-  const record = value as Record<string, unknown>
-  const total = normalizeTokenUsageBreakdown(record.total)
-  const last = normalizeTokenUsageBreakdown(record.last)
-  if (!total || !last) return null
-
-  const modelContextWindow = normalizeStoredTokenCount(record.modelContextWindow)
-  const currentContextTokens = last.totalTokens
-  const remainingContextTokens = typeof modelContextWindow === 'number'
-    ? Math.max(modelContextWindow - currentContextTokens, 0)
-    : null
-  const remainingContextPercent = typeof modelContextWindow === 'number' && modelContextWindow > 0
-    ? clamp(Math.round((remainingContextTokens ?? 0) / modelContextWindow * 100), 0, 100)
-    : null
-
-  return {
-    total,
-    last,
-    modelContextWindow,
-    currentContextTokens,
-    remainingContextTokens,
-    remainingContextPercent,
-  }
-}
-
-function loadThreadTokenUsageMap(): Record<string, UiThreadTokenUsage> {
-  if (typeof window === 'undefined') return {}
-
-  try {
-    const raw = window.localStorage.getItem(THREAD_TOKEN_USAGE_STORAGE_KEY)
-    if (!raw) return {}
-
-    const parsed = JSON.parse(raw) as unknown
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return {}
-
-    const normalizedMap: Record<string, UiThreadTokenUsage> = {}
-    for (const [threadId, usage] of Object.entries(parsed as Record<string, unknown>)) {
-      if (!threadId) continue
-      const normalizedUsage = normalizeThreadTokenUsage(usage)
-      if (normalizedUsage) {
-        normalizedMap[threadId] = normalizedUsage
-      }
-    }
-    return normalizedMap
-  } catch {
-    return {}
-  }
-}
-
-function saveThreadTokenUsageMap(state: Record<string, UiThreadTokenUsage>): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(THREAD_TOKEN_USAGE_STORAGE_KEY, JSON.stringify(state))
-}
-
-function loadThreadTerminalOpenMap(): Record<string, boolean> {
-  if (typeof window === 'undefined') return {}
-
-  try {
-    const raw = window.localStorage.getItem(THREAD_TERMINAL_OPEN_STORAGE_KEY)
-    if (!raw) return {}
-
-    const parsed = JSON.parse(raw) as unknown
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return {}
-
-    const normalizedMap: Record<string, boolean> = {}
-    for (const [threadId, isOpen] of Object.entries(parsed as Record<string, unknown>)) {
-      if (threadId && typeof isOpen === 'boolean') {
-        normalizedMap[threadId] = isOpen
-      }
-    }
-    return normalizedMap
-  } catch {
-    return {}
-  }
-}
-
-function saveThreadTerminalOpenMap(state: Record<string, boolean>): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(THREAD_TERMINAL_OPEN_STORAGE_KEY, JSON.stringify(state))
-}
-
-function loadSelectedThreadId(): string {
-  if (typeof window === 'undefined') return ''
-  const raw = window.localStorage.getItem(SELECTED_THREAD_STORAGE_KEY)
-  return raw ?? ''
-}
-
-function saveSelectedThreadId(threadId: string): void {
-  if (typeof window === 'undefined') return
-  if (!threadId) {
-    window.localStorage.removeItem(SELECTED_THREAD_STORAGE_KEY)
-    return
-  }
-  window.localStorage.setItem(SELECTED_THREAD_STORAGE_KEY, threadId)
-}
-
-function loadProjectOrder(): string[] {
-  if (typeof window === 'undefined') return []
-
-  try {
-    const raw = window.localStorage.getItem(PROJECT_ORDER_STORAGE_KEY)
-    if (!raw) return []
-
-    const parsed = JSON.parse(raw) as unknown
-    if (!Array.isArray(parsed)) return []
-    const order: string[] = []
-    for (const item of parsed) {
-      if (typeof item !== 'string' || item.length === 0) continue
-      const normalizedItem = toProjectName(item)
-      if (normalizedItem.length > 0 && !order.includes(normalizedItem)) {
-        order.push(normalizedItem)
-      }
-    }
-    return order
-  } catch {
-    return []
-  }
-}
-
-function saveProjectOrder(order: string[]): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(PROJECT_ORDER_STORAGE_KEY, JSON.stringify(order))
-}
-
-function loadProjectDisplayNames(): Record<string, string> {
-  if (typeof window === 'undefined') return {}
-
-  try {
-    const raw = window.localStorage.getItem(PROJECT_DISPLAY_NAME_STORAGE_KEY)
-    if (!raw) return {}
-
-    const parsed = JSON.parse(raw) as unknown
-    if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) return {}
-
-    const displayNames: Record<string, string> = {}
-    for (const [projectName, displayName] of Object.entries(parsed as Record<string, unknown>)) {
-      const normalizedProjectName = typeof projectName === 'string' ? toProjectName(projectName) : ''
-      if (normalizedProjectName.length > 0 && typeof displayName === 'string') {
-        displayNames[normalizedProjectName] = displayName
-      }
-    }
-    return displayNames
-  } catch {
-    return {}
-  }
-}
-
-function saveProjectDisplayNames(displayNames: Record<string, string>): void {
-  if (typeof window === 'undefined') return
-  window.localStorage.setItem(PROJECT_DISPLAY_NAME_STORAGE_KEY, JSON.stringify(displayNames))
-}
-
-function mergeProjectOrder(previousOrder: string[], incomingGroups: UiProjectGroup[]): string[] {
-  const nextOrder: string[] = []
-
-  for (const projectName of previousOrder) {
-    if (!nextOrder.includes(projectName)) {
-      nextOrder.push(projectName)
-    }
-  }
-
-  for (const group of incomingGroups) {
-    if (!nextOrder.includes(group.projectName)) {
-      nextOrder.push(group.projectName)
-    }
-  }
-
-  return areStringArraysEqual(previousOrder, nextOrder) ? previousOrder : nextOrder
-}
-
-function orderGroupsByProjectOrder(incoming: UiProjectGroup[], projectOrder: string[]): UiProjectGroup[] {
-  const incomingByName = new Map(incoming.map((group) => [group.projectName, group]))
-  const ordered: UiProjectGroup[] = projectOrder
-    .map((projectName) => incomingByName.get(projectName) ?? null)
-    .filter((group): group is UiProjectGroup => group !== null && group.threads.length > 0)
-
-  for (const group of incoming) {
-    if (!projectOrder.includes(group.projectName)) {
-      ordered.push(group)
-    }
-  }
-
-  return ordered
-}
-
-function areStringArraysEqual(first?: string[], second?: string[]): boolean {
-  const left = Array.isArray(first) ? first : []
-  const right = Array.isArray(second) ? second : []
-  if (left.length !== right.length) return false
-  for (let index = 0; index < left.length; index += 1) {
-    if (left[index] !== right[index]) return false
-  }
-  return true
-}
-
-function reorderStringArray(items: string[], fromIndex: number, toIndex: number): string[] {
-  if (fromIndex < 0 || fromIndex >= items.length || toIndex < 0 || toIndex >= items.length) {
-    return items
-  }
-
-  if (fromIndex === toIndex) {
-    return items
-  }
-
-  const next = [...items]
-  const [moved] = next.splice(fromIndex, 1)
-  next.splice(toIndex, 0, moved)
-  return next
-}
-
-function areCommandExecutionsEqual(first?: CommandExecutionData, second?: CommandExecutionData): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  return first.status === second.status && first.aggregatedOutput === second.aggregatedOutput && first.exitCode === second.exitCode
-}
-
-function arePlanStepsEqual(first: UiPlanStep[] = [], second: UiPlanStep[] = []): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index]?.step !== second[index]?.step || first[index]?.status !== second[index]?.status) {
-      return false
-    }
-  }
-  return true
-}
-
-function arePlanDataEqual(first?: UiPlanData, second?: UiPlanData): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  return (
-    first.explanation === second.explanation &&
-    first.isStreaming === second.isStreaming &&
-    arePlanStepsEqual(first.steps, second.steps)
-  )
-}
-
-function isUnsupportedChatGptModelError(error: unknown): boolean {
-  if (!(error instanceof Error)) return false
-  const message = error.message.toLowerCase()
-  return (
-    message.includes('not supported when using codex with a chatgpt account') ||
-    message.includes('model is not supported')
-  )
-}
-
-function areMessageFieldsEqual(first: UiMessage, second: UiMessage): boolean {
-  return (
-    first.id === second.id &&
-    first.role === second.role &&
-    first.text === second.text &&
-    areStringArraysEqual(first.images, second.images) &&
-    areUiFileChangesEqual(first.fileChanges, second.fileChanges) &&
-    first.fileChangeStatus === second.fileChangeStatus &&
-    first.messageType === second.messageType &&
-    first.rawPayload === second.rawPayload &&
-    first.isUnhandled === second.isUnhandled &&
-    areCommandExecutionsEqual(first.commandExecution, second.commandExecution) &&
-    arePlanDataEqual(first.plan, second.plan) &&
-    first.turnId === second.turnId &&
-    first.turnIndex === second.turnIndex &&
-    first.isAutomationRun === second.isAutomationRun &&
-    first.automationDisplayName === second.automationDisplayName
-  )
-}
-
-function areMessageArraysEqual(first: UiMessage[], second: UiMessage[]): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index] !== second[index]) return false
-  }
-  return true
-}
-
-function mergeMessages(
-  previous: UiMessage[],
-  incoming: UiMessage[],
-  options: { preserveMissing?: boolean } = {},
-): UiMessage[] {
-  const previousById = new Map(previous.map((message) => [message.id, message]))
-  const incomingById = new Map(incoming.map((message) => [message.id, message]))
-
-  const mergedIncoming = incoming.map((incomingMessage) => {
-    const previousMessage = previousById.get(incomingMessage.id)
-    if (previousMessage && areMessageFieldsEqual(previousMessage, incomingMessage)) {
-      return previousMessage
-    }
-    return incomingMessage
-  })
-
-  if (options.preserveMissing !== true) {
-    return areMessageArraysEqual(previous, mergedIncoming) ? previous : mergedIncoming
-  }
-
-  const mergedFromPrevious = previous.map((previousMessage) => {
-    const nextMessage = incomingById.get(previousMessage.id)
-    if (!nextMessage) {
-      return previousMessage
-    }
-    if (areMessageFieldsEqual(previousMessage, nextMessage)) {
-      return previousMessage
-    }
-    return nextMessage
-  })
-
-  const previousIdSet = new Set(previous.map((message) => message.id))
-  const appended = mergedIncoming.filter((message) => !previousIdSet.has(message.id))
-  const merged = [...mergedFromPrevious, ...appended]
-
-  return areMessageArraysEqual(previous, merged) ? previous : merged
-}
-
-function areUiFileChangesEqual(first?: UiFileChange[], second?: UiFileChange[]): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    const firstChange = first[index]
-    const secondChange = second[index]
-    if (
-      firstChange.path !== secondChange.path ||
-      firstChange.operation !== secondChange.operation ||
-      firstChange.movedToPath !== secondChange.movedToPath ||
-      firstChange.diff !== secondChange.diff ||
-      firstChange.addedLineCount !== secondChange.addedLineCount ||
-      firstChange.removedLineCount !== secondChange.removedLineCount
-    ) {
-      return false
-    }
-  }
-  return true
-}
-
-function normalizeMessageText(value: string): string {
-  return value.replace(/\s+/gu, ' ').trim()
-}
-
-function removeRedundantLiveAgentMessages(previous: UiMessage[], incoming: UiMessage[]): UiMessage[] {
-  const incomingMessageIds = new Set(incoming.map((message) => message.id))
-  const incomingAssistantTexts = new Set(
-    incoming
-      .filter((message) => message.role === 'assistant')
-      .map((message) => normalizeMessageText(message.text))
-      .filter((text) => text.length > 0),
-  )
-
-  if (incomingAssistantTexts.size === 0) {
-    return previous
-  }
-
-  const next = previous.filter((message) => {
-    if (message.messageType !== 'agentMessage.live') return true
-    if (incomingMessageIds.has(message.id)) return false
-    const normalized = normalizeMessageText(message.text)
-    if (normalized.length === 0) return false
-    return !incomingAssistantTexts.has(normalized)
-  })
-
-  return next.length === previous.length ? previous : next
-}
-
-function removePersistedLiveMessages(previous: UiMessage[], incoming: UiMessage[]): UiMessage[] {
-  const incomingIds = new Set(incoming.map((message) => message.id))
-  const next = previous.filter((message) => !incomingIds.has(message.id))
-  return next.length === previous.length ? previous : next
-}
-
-function upsertMessage(previous: UiMessage[], nextMessage: UiMessage): UiMessage[] {
-  const existingIndex = previous.findIndex((message) => message.id === nextMessage.id)
-  if (existingIndex < 0) {
-    return [...previous, nextMessage]
-  }
-
-  const existing = previous[existingIndex]
-  if (areMessageFieldsEqual(existing, nextMessage)) {
-    return previous
-  }
-
-  const next = [...previous]
-  next.splice(existingIndex, 1, nextMessage)
-  return next
-}
-
-type TurnSummaryState = {
-  turnId: string
-  durationMs: number
-}
-
-type TurnActivityState = {
-  label: string
-  details: string[]
-}
-
-type TurnErrorState = {
-  message: string
-  transient: boolean
-}
-
-type TurnStartedInfo = {
-  threadId: string
-  turnId: string
-  startedAtMs: number
-}
-
-type TurnCompletedInfo = {
-  threadId: string
-  turnId: string
-  completedAtMs: number
-  startedAtMs?: number
-}
-
-const WORKED_MESSAGE_TYPE = 'worked'
-
-function parseIsoTimestamp(value: string): number | null {
-  if (!value) return null
-  const ms = new Date(value).getTime()
-  return Number.isNaN(ms) ? null : ms
-}
-
-function formatTurnDuration(durationMs: number): string {
-  if (!Number.isFinite(durationMs) || durationMs <= 0) {
-    return '<1s'
-  }
-
-  const totalSeconds = Math.max(1, Math.round(durationMs / 1000))
-  const hours = Math.floor(totalSeconds / 3600)
-  const minutes = Math.floor((totalSeconds % 3600) / 60)
-  const seconds = totalSeconds % 60
-  const parts: string[] = []
-
-  if (hours > 0) {
-    parts.push(`${hours}h`)
-  }
-
-  if (minutes > 0 || hours > 0) {
-    parts.push(`${minutes}m`)
-  }
-
-  const displaySeconds = seconds > 0 || parts.length === 0 ? seconds : 0
-  parts.push(`${displaySeconds}s`)
-  return parts.join(' ')
-}
-
-function areTurnSummariesEqual(first?: TurnSummaryState, second?: TurnSummaryState): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  return first.turnId === second.turnId && first.durationMs === second.durationMs
-}
-
-function areTurnActivitiesEqual(first?: TurnActivityState, second?: TurnActivityState): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  if (first.label !== second.label) return false
-  if (first.details.length !== second.details.length) return false
-  for (let index = 0; index < first.details.length; index += 1) {
-    if (first.details[index] !== second.details[index]) return false
-  }
-  return true
-}
-
-function buildTurnSummaryMessage(summary: TurnSummaryState): UiMessage {
-  return {
-    id: `turn-summary:${summary.turnId}`,
-    role: 'system',
-    text: `Worked for ${formatTurnDuration(summary.durationMs)}`,
-    messageType: WORKED_MESSAGE_TYPE,
-    turnId: summary.turnId,
-  }
-}
-
-function findLastAssistantMessageIndex(messages: UiMessage[]): number {
-  for (let index = messages.length - 1; index >= 0; index -= 1) {
-    if (messages[index].role === 'assistant') {
-      return index
-    }
-  }
-  return -1
-}
-
-function insertTurnSummaryMessage(messages: UiMessage[], summary: TurnSummaryState): UiMessage[] {
-  const summaryMessage = buildTurnSummaryMessage(summary)
-  const sanitizedMessages = messages.filter((message) => message.messageType !== WORKED_MESSAGE_TYPE)
-  const insertIndex = findLastAssistantMessageIndex(sanitizedMessages)
-  if (insertIndex < 0) {
-    return [...sanitizedMessages, summaryMessage]
-  }
-  const next = [...sanitizedMessages]
-  next.splice(insertIndex, 0, summaryMessage)
-  return next
-}
-
-function omitKey<TValue>(record: Record<string, TValue>, key: string): Record<string, TValue> {
-  if (!(key in record)) return record
-  const next = { ...record }
-  delete next[key]
-  return next
-}
-
-function omitKeys<TValue>(record: Record<string, TValue>, keys: Set<string>): Record<string, TValue> {
-  if (keys.size === 0) return record
-  let changed = false
-  const next: Record<string, TValue> = {}
-  for (const [key, value] of Object.entries(record)) {
-    if (keys.has(key)) {
-      changed = true
-      continue
-    }
-    next[key] = value
-  }
-  return changed ? next : record
-}
-
-function areThreadFieldsEqual(first: UiThread, second: UiThread): boolean {
-  return (
-    first.id === second.id &&
-    first.title === second.title &&
-    first.projectName === second.projectName &&
-    first.cwd === second.cwd &&
-    first.createdAtIso === second.createdAtIso &&
-    first.updatedAtIso === second.updatedAtIso &&
-    first.preview === second.preview &&
-    first.unread === second.unread &&
-    first.inProgress === second.inProgress &&
-    first.pendingRequestState === second.pendingRequestState
-  )
-}
-
-function areThreadArraysEqual(first: UiThread[], second: UiThread[]): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index] !== second[index]) return false
-  }
-  return true
-}
-
-function areGroupArraysEqual(first: UiProjectGroup[], second: UiProjectGroup[]): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index] !== second[index]) return false
-  }
-  return true
-}
-
-function pruneThreadStateMap<T>(stateMap: Record<string, T>, threadIds: Set<string>): Record<string, T> {
-  const nextEntries = Object.entries(stateMap).filter(([threadId]) => threadIds.has(threadId))
-  if (nextEntries.length === Object.keys(stateMap).length) {
-    return stateMap
-  }
-  return Object.fromEntries(nextEntries) as Record<string, T>
-}
-
-function mergeThreadGroups(
-  previous: UiProjectGroup[],
-  incoming: UiProjectGroup[],
-): UiProjectGroup[] {
-  const previousGroupsByName = new Map(previous.map((group) => [group.projectName, group]))
-  const mergedGroups: UiProjectGroup[] = incoming.map((incomingGroup) => {
-    const previousGroup = previousGroupsByName.get(incomingGroup.projectName)
-    const previousThreadsById = new Map(previousGroup?.threads.map((thread) => [thread.id, thread]) ?? [])
-
-    const mergedThreads = incomingGroup.threads.map((incomingThread) => {
-      const previousThread = previousThreadsById.get(incomingThread.id)
-      if (previousThread && areThreadFieldsEqual(previousThread, incomingThread)) {
-        return previousThread
-      }
-      return incomingThread
-    })
-
-    if (
-      previousGroup &&
-      previousGroup.projectName === incomingGroup.projectName &&
-      areThreadArraysEqual(previousGroup.threads, mergedThreads)
-    ) {
-      return previousGroup
-    }
-
-    return {
-      projectName: incomingGroup.projectName,
-      threads: mergedThreads,
-    }
-  })
-
-  return areGroupArraysEqual(previous, mergedGroups) ? previous : mergedGroups
-}
-
-function mergeIncomingWithLocalInProgressThreads(
-  previous: UiProjectGroup[],
-  incoming: UiProjectGroup[],
-  inProgressById: Record<string, boolean>,
-): UiProjectGroup[] {
-  const incomingThreadIds = new Set(flattenThreads(incoming).map((thread) => thread.id))
-  const localInProgressThreads = flattenThreads(previous).filter(
-    (thread) => inProgressById[thread.id] === true && !incomingThreadIds.has(thread.id),
-  )
-
-  if (localInProgressThreads.length === 0) {
-    return incoming
-  }
-
-  const incomingByProjectName = new Map(incoming.map((group) => [group.projectName, group]))
-  const merged: UiProjectGroup[] = incoming.map((group) => ({
-    projectName: group.projectName,
-    threads: [...group.threads],
-  }))
-
-  for (const thread of localInProgressThreads) {
-    const existingGroup = incomingByProjectName.get(thread.projectName)
-    if (existingGroup) {
-      const mergedGroupIndex = merged.findIndex((group) => group.projectName === thread.projectName)
-      if (mergedGroupIndex >= 0) {
-        merged[mergedGroupIndex] = {
-          projectName: merged[mergedGroupIndex].projectName,
-          threads: [thread, ...merged[mergedGroupIndex].threads],
-        }
-      }
-      continue
-    }
-
-    merged.push({
-      projectName: thread.projectName,
-      threads: [thread],
-    })
-  }
-
-  return merged
-}
-
-function toProjectNameFromWorkspaceRoot(value: string): string {
-  return toProjectName(value)
-}
-
-function toOptimisticThreadTitle(message: string): string {
-  const firstLine = message
-    .split('\n')
-    .map((line) => line.trim())
-    .find((line) => line.length > 0)
-
-  if (!firstLine) return 'Untitled thread'
-  return firstLine.slice(0, 80)
-}
-
-function toForkedThreadTitle(title: string): string {
-  const normalizedTitle = title.trim() || 'Untitled thread'
-  return /^fork:\s+/iu.test(normalizedTitle) ? normalizedTitle : `Fork: ${normalizedTitle}`
-}
 
 export function useDesktopState() {
   const projectGroups = ref<UiProjectGroup[]>([])
