+++ src/features/thread-state/merge.ts
@@ -1,0 +1,205 @@
+import type {
+  UiProjectGroup,
+  UiThread,
+} from '../../types/codex'
+import { toProjectName } from '../../pathUtils.js'
+
+export function flattenThreads(groups: UiProjectGroup[]): UiThread[] {
+  return groups.flatMap((group) => group.threads)
+}
+
+export function mergeProjectOrder(previousOrder: string[], incomingGroups: UiProjectGroup[]): string[] {
+  const nextOrder: string[] = []
+  for (const projectName of previousOrder) {
+    if (!nextOrder.includes(projectName)) {
+      nextOrder.push(projectName)
+    }
+  }
+  for (const group of incomingGroups) {
+    if (!nextOrder.includes(group.projectName)) {
+      nextOrder.push(group.projectName)
+    }
+  }
+  return areStringArraysEqual(previousOrder, nextOrder) ? previousOrder : nextOrder
+}
+
+export function orderGroupsByProjectOrder(incoming: UiProjectGroup[], projectOrder: string[]): UiProjectGroup[] {
+  const incomingByName = new Map(incoming.map((group) => [group.projectName, group]))
+  const ordered: UiProjectGroup[] = projectOrder
+    .map((projectName) => incomingByName.get(projectName) ?? null)
+    .filter((group): group is UiProjectGroup => group !== null && group.threads.length > 0)
+  for (const group of incoming) {
+    if (!projectOrder.includes(group.projectName)) {
+      ordered.push(group)
+    }
+  }
+  return ordered
+}
+
+export function areStringArraysEqual(first?: string[], second?: string[]): boolean {
+  const left = Array.isArray(first) ? first : []
+  const right = Array.isArray(second) ? second : []
+  if (left.length !== right.length) return false
+  for (let index = 0; index < left.length; index += 1) {
+    if (left[index] !== right[index]) return false
+  }
+  return true
+}
+
+export function reorderStringArray(items: string[], fromIndex: number, toIndex: number): string[] {
+  if (fromIndex < 0 || fromIndex >= items.length || toIndex < 0 || toIndex >= items.length) {
+    return items
+  }
+  if (fromIndex === toIndex) {
+    return items
+  }
+  const next = [...items]
+  const [moved] = next.splice(fromIndex, 1)
+  next.splice(toIndex, 0, moved)
+  return next
+}
+
+export function omitKey<TValue>(record: Record<string, TValue>, key: string): Record<string, TValue> {
+  if (!(key in record)) return record
+  const next = { ...record }
+  delete next[key]
+  return next
+}
+
+export function omitKeys<TValue>(record: Record<string, TValue>, keys: Set<string>): Record<string, TValue> {
+  if (keys.size === 0) return record
+  let changed = false
+  const next: Record<string, TValue> = {}
+  for (const [key, value] of Object.entries(record)) {
+    if (keys.has(key)) {
+      changed = true
+      continue
+    }
+    next[key] = value
+  }
+  return changed ? next : record
+}
+
+export function areThreadFieldsEqual(first: UiThread, second: UiThread): boolean {
+  return (
+    first.id === second.id &&
+    first.title === second.title &&
+    first.projectName === second.projectName &&
+    first.cwd === second.cwd &&
+    first.createdAtIso === second.createdAtIso &&
+    first.updatedAtIso === second.updatedAtIso &&
+    first.preview === second.preview &&
+    first.unread === second.unread &&
+    first.inProgress === second.inProgress &&
+    first.pendingRequestState === second.pendingRequestState
+  )
+}
+
+export function areThreadArraysEqual(first: UiThread[], second: UiThread[]): boolean {
+  if (first.length !== second.length) return false
+  for (let index = 0; index < first.length; index += 1) {
+    if (first[index] !== second[index]) return false
+  }
+  return true
+}
+
+export function areGroupArraysEqual(first: UiProjectGroup[], second: UiProjectGroup[]): boolean {
+  if (first.length !== second.length) return false
+  for (let index = 0; index < first.length; index += 1) {
+    if (first[index] !== second[index]) return false
+  }
+  return true
+}
+
+export function pruneThreadStateMap<T>(stateMap: Record<string, T>, threadIds: Set<string>): Record<string, T> {
+  const nextEntries = Object.entries(stateMap).filter(([threadId]) => threadIds.has(threadId))
+  if (nextEntries.length === Object.keys(stateMap).length) {
+    return stateMap
+  }
+  return Object.fromEntries(nextEntries) as Record<string, T>
+}
+
+export function mergeThreadGroups(
+  previous: UiProjectGroup[],
+  incoming: UiProjectGroup[],
+): UiProjectGroup[] {
+  const previousGroupsByName = new Map(previous.map((group) => [group.projectName, group]))
+  const mergedGroups: UiProjectGroup[] = incoming.map((incomingGroup) => {
+    const previousGroup = previousGroupsByName.get(incomingGroup.projectName)
+    const previousThreadsById = new Map(previousGroup?.threads.map((thread) => [thread.id, thread]) ?? [])
+    const mergedThreads = incomingGroup.threads.map((incomingThread) => {
+      const previousThread = previousThreadsById.get(incomingThread.id)
+      if (previousThread && areThreadFieldsEqual(previousThread, incomingThread)) {
+        return previousThread
+      }
+      return incomingThread
+    })
+    if (
+      previousGroup &&
+      previousGroup.projectName === incomingGroup.projectName &&
+      areThreadArraysEqual(previousGroup.threads, mergedThreads)
+    ) {
+      return previousGroup
+    }
+    return {
+      projectName: incomingGroup.projectName,
+      threads: mergedThreads,
+    }
+  })
+  return areGroupArraysEqual(previous, mergedGroups) ? previous : mergedGroups
+}
+
+export function mergeIncomingWithLocalInProgressThreads(
+  previous: UiProjectGroup[],
+  incoming: UiProjectGroup[],
+  inProgressById: Record<string, boolean>,
+): UiProjectGroup[] {
+  const incomingThreadIds = new Set(flattenThreads(incoming).map((thread) => thread.id))
+  const localInProgressThreads = flattenThreads(previous).filter(
+    (thread) => inProgressById[thread.id] === true && !incomingThreadIds.has(thread.id),
+  )
+  if (localInProgressThreads.length === 0) {
+    return incoming
+  }
+  const incomingByProjectName = new Map(incoming.map((group) => [group.projectName, group]))
+  const merged: UiProjectGroup[] = incoming.map((group) => ({
+    projectName: group.projectName,
+    threads: [...group.threads],
+  }))
+  for (const thread of localInProgressThreads) {
+    const existingGroup = incomingByProjectName.get(thread.projectName)
+    if (existingGroup) {
+      const mergedGroupIndex = merged.findIndex((group) => group.projectName === thread.projectName)
+      if (mergedGroupIndex >= 0) {
+        merged[mergedGroupIndex] = {
+          projectName: merged[mergedGroupIndex].projectName,
+          threads: [thread, ...merged[mergedGroupIndex].threads],
+        }
+      }
+      continue
+    }
+    merged.push({
+      projectName: thread.projectName,
+      threads: [thread],
+    })
+  }
+  return merged
+}
+
+export function toProjectNameFromWorkspaceRoot(value: string): string {
+  return toProjectName(value)
+}
+
+export function toOptimisticThreadTitle(message: string): string {
+  const firstLine = message
+    .split('\n')
+    .map((line) => line.trim())
+    .find((line) => line.length > 0)
+  if (!firstLine) return 'Untitled thread'
+  return firstLine.slice(0, 80)
+}
+
+export function toForkedThreadTitle(title: string): string {
+  const normalizedTitle = title.trim() || 'Untitled thread'
+  return /^fork:\s+/iu.test(normalizedTitle) ? normalizedTitle : `Fork: ${normalizedTitle}`
+}
