+++ src/features/thread-state/messages.ts
@@ -1,0 +1,284 @@
+import type {
+  CommandExecutionData,
+  UiFileChange,
+  UiMessage,
+  UiPlanData,
+  UiPlanStep,
+  UiThreadTokenUsage,
+} from '../../types/codex'
+
+export type TurnSummaryState = {
+  turnId: string
+  durationMs: number
+}
+
+export type TurnActivityState = {
+  label: string
+  details: string[]
+}
+
+export type TurnErrorState = {
+  message: string
+  transient: boolean
+}
+
+export type TurnStartedInfo = {
+  threadId: string
+  turnId: string
+  startedAtMs: number
+}
+
+export type TurnCompletedInfo = {
+  threadId: string
+  turnId: string
+  completedAtMs: number
+  startedAtMs?: number
+}
+
+const WORKED_MESSAGE_TYPE = 'worked'
+
+export function parseIsoTimestamp(value: string): number | null {
+  if (!value) return null
+  const ms = new Date(value).getTime()
+  return Number.isNaN(ms) ? null : ms
+}
+
+export function formatTurnDuration(durationMs: number): string {
+  if (!Number.isFinite(durationMs) || durationMs <= 0) {
+    return '<1s'
+  }
+  const totalSeconds = Math.max(1, Math.round(durationMs / 1000))
+  const hours = Math.floor(totalSeconds / 3600)
+  const minutes = Math.floor((totalSeconds % 3600) / 60)
+  const seconds = totalSeconds % 60
+  const parts: string[] = []
+  if (hours > 0) {
+    parts.push(`${hours}h`)
+  }
+  if (minutes > 0 || hours > 0) {
+    parts.push(`${minutes}m`)
+  }
+  const displaySeconds = seconds > 0 || parts.length === 0 ? seconds : 0
+  parts.push(`${displaySeconds}s`)
+  return parts.join(' ')
+}
+
+export function areTurnSummariesEqual(first?: TurnSummaryState, second?: TurnSummaryState): boolean {
+  if (!first && !second) return true
+  if (!first || !second) return false
+  return first.turnId === second.turnId && first.durationMs === second.durationMs
+}
+
+export function areTurnActivitiesEqual(first?: TurnActivityState, second?: TurnActivityState): boolean {
+  if (!first && !second) return true
+  if (!first || !second) return false
+  if (first.label !== second.label) return false
+  if (first.details.length !== second.details.length) return false
+  for (let index = 0; index < first.details.length; index += 1) {
+    if (first.details[index] !== second.details[index]) return false
+  }
+  return true
+}
+
+export function buildTurnSummaryMessage(summary: TurnSummaryState): UiMessage {
+  return {
+    id: `turn-summary:${summary.turnId}`,
+    role: 'system',
+    text: `Worked for ${formatTurnDuration(summary.durationMs)}`,
+    messageType: WORKED_MESSAGE_TYPE,
+    turnId: summary.turnId,
+  }
+}
+
+export function findLastAssistantMessageIndex(messages: UiMessage[]): number {
+  for (let index = messages.length - 1; index >= 0; index -= 1) {
+    if (messages[index].role === 'assistant') {
+      return index
+    }
+  }
+  return -1
+}
+
+export function insertTurnSummaryMessage(messages: UiMessage[], summary: TurnSummaryState): UiMessage[] {
+  const summaryMessage = buildTurnSummaryMessage(summary)
+  const sanitizedMessages = messages.filter((message) => message.messageType !== WORKED_MESSAGE_TYPE)
+  const insertIndex = findLastAssistantMessageIndex(sanitizedMessages)
+  if (insertIndex < 0) {
+    return [...sanitizedMessages, summaryMessage]
+  }
+  const next = [...sanitizedMessages]
+  next.splice(insertIndex, 0, summaryMessage)
+  return next
+}
+
+export function areCommandExecutionsEqual(first?: CommandExecutionData, second?: CommandExecutionData): boolean {
+  if (!first && !second) return true
+  if (!first || !second) return false
+  return first.status === second.status && first.aggregatedOutput === second.aggregatedOutput && first.exitCode === second.exitCode
+}
+
+export function arePlanStepsEqual(first: UiPlanStep[] = [], second: UiPlanStep[] = []): boolean {
+  if (first.length !== second.length) return false
+  for (let index = 0; index < first.length; index += 1) {
+    if (first[index]?.step !== second[index]?.step || first[index]?.status !== second[index]?.status) {
+      return false
+    }
+  }
+  return true
+}
+
+export function arePlanDataEqual(first?: UiPlanData, second?: UiPlanData): boolean {
+  if (!first && !second) return true
+  if (!first || !second) return false
+  return (
+    first.explanation === second.explanation &&
+    first.isStreaming === second.isStreaming &&
+    arePlanStepsEqual(first.steps, second.steps)
+  )
+}
+
+export function isUnsupportedChatGptModelError(error: unknown): boolean {
+  if (!(error instanceof Error)) return false
+  const message = error.message.toLowerCase()
+  return (
+    message.includes('not supported when using codex with a chatgpt account') ||
+    message.includes('model is not supported')
+  )
+}
+
+export function areUiFileChangesEqual(first?: UiFileChange[], second?: UiFileChange[]): boolean {
+  if (!first && !second) return true
+  if (!first || !second) return false
+  if (first.length !== second.length) return false
+  for (let index = 0; index < first.length; index += 1) {
+    const firstChange = first[index]
+    const secondChange = second[index]
+    if (
+      firstChange.path !== secondChange.path ||
+      firstChange.operation !== secondChange.operation ||
+      firstChange.movedToPath !== secondChange.movedToPath ||
+      firstChange.diff !== secondChange.diff ||
+      firstChange.addedLineCount !== secondChange.addedLineCount ||
+      firstChange.removedLineCount !== secondChange.removedLineCount
+    ) {
+      return false
+    }
+  }
+  return true
+}
+
+export function areMessageFieldsEqual(first: UiMessage, second: UiMessage): boolean {
+  return (
+    first.id === second.id &&
+    first.role === second.role &&
+    first.text === second.text &&
+    areStringArraysEqual(first.images, second.images) &&
+    areUiFileChangesEqual(first.fileChanges, second.fileChanges) &&
+    first.fileChangeStatus === second.fileChangeStatus &&
+    first.messageType === second.messageType &&
+    first.rawPayload === second.rawPayload &&
+    first.isUnhandled === second.isUnhandled &&
+    areCommandExecutionsEqual(first.commandExecution, second.commandExecution) &&
+    arePlanDataEqual(first.plan, second.plan) &&
+    first.turnId === second.turnId &&
+    first.turnIndex === second.turnIndex &&
+    first.isAutomationRun === second.isAutomationRun &&
+    first.automationDisplayName === second.automationDisplayName
+  )
+}
+
+export function areMessageArraysEqual(first: UiMessage[], second: UiMessage[]): boolean {
+  if (first.length !== second.length) return false
+  for (let index = 0; index < first.length; index += 1) {
+    if (first[index] !== second[index]) return false
+  }
+  return true
+}
+
+export function mergeMessages(
+  previous: UiMessage[],
+  incoming: UiMessage[],
+  options: { preserveMissing?: boolean } = {},
+): UiMessage[] {
+  const previousById = new Map(previous.map((message) => [message.id, message]))
+  const incomingById = new Map(incoming.map((message) => [message.id, message]))
+  const mergedIncoming = incoming.map((incomingMessage) => {
+    const previousMessage = previousById.get(incomingMessage.id)
+    if (previousMessage && areMessageFieldsEqual(previousMessage, incomingMessage)) {
+      return previousMessage
+    }
+    return incomingMessage
+  })
+  if (options.preserveMissing !== true) {
+    return areMessageArraysEqual(previous, mergedIncoming) ? previous : mergedIncoming
+  }
+  const mergedFromPrevious = previous.map((previousMessage) => {
+    const nextMessage = incomingById.get(previousMessage.id)
+    if (!nextMessage) {
+      return previousMessage
+    }
+    if (areMessageFieldsEqual(previousMessage, nextMessage)) {
+      return previousMessage
+    }
+    return nextMessage
+  })
+  const previousIdSet = new Set(previous.map((message) => message.id))
+  const appended = mergedIncoming.filter((message) => !previousIdSet.has(message.id))
+  const merged = [...mergedFromPrevious, ...appended]
+  return areMessageArraysEqual(previous, merged) ? previous : merged
+}
+
+export function normalizeMessageText(value: string): string {
+  return value.replace(/\s+/gu, ' ').trim()
+}
+
+export function removeRedundantLiveAgentMessages(previous: UiMessage[], incoming: UiMessage[]): UiMessage[] {
+  const incomingMessageIds = new Set(incoming.map((message) => message.id))
+  const incomingAssistantTexts = new Set(
+    incoming
+      .filter((message) => message.role === 'assistant')
+      .map((message) => normalizeMessageText(message.text))
+      .filter((text) => text.length > 0),
+  )
+  if (incomingAssistantTexts.size === 0) {
+    return previous
+  }
+  const next = previous.filter((message) => {
+    if (message.messageType !== 'agentMessage.live') return true
+    if (incomingMessageIds.has(message.id)) return false
+    const normalized = normalizeMessageText(message.text)
+    if (normalized.length === 0) return false
+    return !incomingAssistantTexts.has(normalized)
+  })
+  return next.length === previous.length ? previous : next
+}
+
+export function removePersistedLiveMessages(previous: UiMessage[], incoming: UiMessage[]): UiMessage[] {
+  const incomingIds = new Set(incoming.map((message) => message.id))
+  const next = previous.filter((message) => !incomingIds.has(message.id))
+  return next.length === previous.length ? previous : next
+}
+
+export function upsertMessage(previous: UiMessage[], nextMessage: UiMessage): UiMessage[] {
+  const existingIndex = previous.findIndex((message) => message.id === nextMessage.id)
+  if (existingIndex < 0) {
+    return [...previous, nextMessage]
+  }
+  const existing = previous[existingIndex]
+  if (areMessageFieldsEqual(existing, nextMessage)) {
+    return previous
+  }
+  const next = [...previous]
+  next.splice(existingIndex, 1, nextMessage)
+  return next
+}
+
+function areStringArraysEqual(first?: string[], second?: string[]): boolean {
+  const left = Array.isArray(first) ? first : []
+  const right = Array.isArray(second) ? second : []
+  if (left.length !== right.length) return false
+  for (let index = 0; index < left.length; index += 1) {
+    if (left[index] !== right[index]) return false
+  }
+  return true
+}
